+++ just-swim-app/src/app/(beforeLogin)/_utils/index.ts
@@ -1,2 +1,1 @@
 export { handleSignUp } from './server/handleSignUp';
-export { setTokenInCookies, getTokenInCookies } from './handleTokenInCookies';
