--- just-swim-app/src/app/(beforeLogin)/signup/type/_components/index.ts
@@ -1,1 +1,0 @@
-export * from './layout';
