--- just-swim-app/src/app/(beforeLogin)/signup/type/_components/layout/index.ts
@@ -1,3 +1,0 @@
-export * from './selectTypeFooter';
-export * from './selectTypeHeader';
-export * from './selectTypeSection';
