--- just-swim-app/src/app/(beforeLogin)/signup/type/_components/layout/selectTypeFooter/styles.module.scss
@@ -1,31 +1,0 @@
-@import '@/globals.scss';
-
-.select_type_footer {
-  flex-grow: 2;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  width: 100%;
-
-  .button_wrapper {
-    width: inherit;
-    height: 52px;
-    border: none;
-    border-radius: 10px;
-    background-color: $--gray-300;
-    pointer-events: none;
-
-    .select_button {
-      @include Title3-SB;
-
-      width: 100%;
-      height: 100%;
-      color: $--White;
-    }
-  }
-
-  .active {
-    background-color: $--Primary-blue;
-    pointer-events: all;
-  }
-}
