--- just-swim-app/src/app/(beforeLogin)/signup/type/_components/layout/selectTypeSection/styles.module.scss
@@ -1,73 +1,0 @@
-@import '@/globals.scss';
-
-.select_type_section {
-  flex-grow: 2;
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-  width: 100%;
-
-  .type_button_wrapper {
-    width: 100%;
-    height: 250px;
-    border-radius: 20px;
-    border: none;
-    cursor: pointer;
-
-    .type_button {
-      display: flex;
-      flex-direction: column;
-      justify-content: center;
-      align-items: center;
-      gap: 25px;
-      // background-color: var(--gray-50);
-      background-color: $--gray-50;
-      border: $--gray-200 1px solid;
-      // border: var(--gray-200) 1px solid;
-      width: 100%;
-      height: 100%;
-      border-radius: 8px;
-    }
-
-    .type_button_img {
-      width: 100%;
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-    }
-
-    .type_button_img div {
-      width: 100px;
-      height: 100px;
-      // background-color: var(--White);
-      background-color: $--White;
-    }
-
-    .type_button_info {
-      display: flex;
-      flex-direction: column;
-      gap: 10px;
-
-      h3 {
-        @include Title2-SB;
-
-        color: $--Black;
-      }
-
-      p {
-        @include Body3-R;
-
-        color: $--gray-500;
-      }
-    }
-
-    .active {
-      background-color: $--Primary-blue;
-
-      .type_button_info h3,
-      p {
-        color: $--White;
-      }
-    }
-  }
-}
