+++ just-swim-app/src/app/_components/common/index.ts
@@ -3,3 +3,4 @@ export * from './calendar';
 export * from './carousel';
 export * from './item';
 export * from './modal';
+export * from './urlImage';
