+++ just-swim-app/src/app/_data/index.ts
@@ -2,3 +2,4 @@ export * from './text';
 export * from './color';
 export * from './common';
 export * from './http';
+export * from './routes';
