+++ just-swim-app/src/app/_data/routes.ts
@@ -5,13 +5,13 @@ const SIGNIN_ROUTE = {
   type: '/signup/type',
   profile: '/signup/profile',
   complete: '/signup/complete',
-};
+} as const;
 
 const SCHEDULE_ROUTE = {
   root: '/schedule',
-};
+} as const;
 
 export const ROUTES = {
   ONBOARDING: { ...SIGNIN_ROUTE },
   SCHEDULE: { ...SCHEDULE_ROUTE },
-};
+} as const;
