+++ just-swim-app/src/app/_data/text/common.ts
@@ -1,0 +1,6 @@
+export const COMMON_TEXT = {
+  select: '선택',
+  next: '다음',
+  skip: '건너뛰기',
+  start: '시작하기',
+} as const;
