+++ just-swim-app/src/app/_data/text/index.ts
@@ -1,0 +1,15 @@
+import { COMMON_TEXT } from './common';
+import {
+  SIGNUP_TEXT,
+  SELECT_TEXT,
+  SET_PROFILE_TEXT,
+  SIGNUP_COMPLETE_TEXT,
+} from './beforeLogin';
+
+export const TEXT = {
+  COMMON: { ...COMMON_TEXT },
+  SIGNUP_PAGE: { ...SIGNUP_TEXT },
+  TYPE_SELECT_PAGE: { ...SELECT_TEXT },
+  SET_PROFILE_PAGE: { ...SET_PROFILE_TEXT },
+  SIGNUP_COMPLETE_PAGE: { ...SIGNUP_COMPLETE_TEXT },
+};
