+++ just-swim-app/src/app/_utils/index.ts
@@ -2,4 +2,5 @@ export * from './numberFormat';
 export * from './optimization';
 export * from './randomId';
 export * from './refs';
-export * from './useURLImage';
+export * from '../_components/common/urlImage';
+export * from './handleTokenInCookies';
